% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-model.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.default}
\alias{extract_model.arima}
\alias{extract_model.Arima}
\alias{extract_model.model_fit}
\alias{extract_model.workflow}
\alias{extract_model.Arima_fit_impl}
\alias{extract_model.mdl_ts}
\alias{extract_model.ARIMA}
\alias{extract_model.AR}
\alias{extract_model.garch}
\alias{extract_model.fGARCH}
\title{Extracts available models}
\usage{
extract_model(object, ...)

\method{extract_model}{default}(object, ...)

\method{extract_model}{arima}(object, ...)

\method{extract_model}{Arima}(object, ...)

\method{extract_model}{model_fit}(object, ...)

\method{extract_model}{workflow}(object, ...)

\method{extract_model}{Arima_fit_impl}(object, ...)

\method{extract_model}{mdl_ts}(object, ...)

\method{extract_model}{ARIMA}(object, ...)

\method{extract_model}{AR}(object, ...)

\method{extract_model}{garch}(object, ...)

\method{extract_model}{fGARCH}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{unused.}
}
\description{
Generic function to extract and return objects that have been registered with
\code{rcits} package. For now only \code{arima} and \code{garch} objects are registered.
This function translates functions of class \code{Arima}, \code{Arima_fit_impl}, \code{AR},
\code{ARIMA} and \code{fGARCH} to the corresponding registered methods. This function
also extracts the fitted models from \code{model_fit}, \code{workflow} and \code{mdl_ts}
objects.
}
