% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-smpspl_grid.R
\name{features.smpspl_grid}
\alias{features.smpspl_grid}
\title{Features for sample splitting}
\usage{
\method{features}{smpspl_grid}(.tbl, .var, .var_nest1, .var_nest2, features, ...)
}
\arguments{
\item{.tbl}{A dataset}

\item{.var}{The variable to compute features on.}

\item{.var_nest1}{The variable \code{.var} is nested in.}

\item{.var_nest2}{The variable \code{.var_nest1} is nested in. This is a column
of \code{.tbl}.}

\item{features}{A list of functions (or lambda expressions) for the features to compute. \code{\link[fabletools:feature_set]{feature_set()}} is a useful helper for building sets of features.}

\item{...}{Additional arguments to be passed to each feature. These arguments will only be passed to features which use it in their formal arguments (\code{\link[base:formals]{base::formals()}}), and not via their \code{...}. While passing \code{na.rm = TRUE} to \code{\link[stats:cor]{stats::var()}} will work, it will not for \code{\link[base:mean]{base::mean()}} as its formals are \code{x} and \code{...}. To more precisely pass inputs to each function, you should use lambdas in the list of features (\code{~ mean(., na.rm = TRUE)}).}
}
\value{
An \code{smpspl_grid_features} object.
}
\description{
Features for sample splitting
}
\examples{
library(fable)
library(smpspl)
data <-
    tsibble::tsibble(x = rnorm(100), date = Sys.Date() + 0:99, index = date)
# Consider an AR(1) model
o <-
    ARIMA(x ~ pdq(1, 0, 0) + PDQ(0, 0, 0)) |>
    smpspl_grid(data, 2, 2)

# Calculate lanyard features
my_acf <-
    function(x) {
        data.frame(t = x, e = 0) |>
            lanyard::acf_metric(t, e) |>
            generics::tidy()
    }

o |>
    features(.resid, .subresid, .assessment, features = my_acf)
}
