% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-smpspl_grid.R
\name{features_at.smpspl_grid}
\alias{features_at.smpspl_grid}
\title{Features for sample splitting}
\usage{
\method{features_at}{smpspl_grid}(.tbl, .vars, .var_nest1, .var_nest2, features, ...)
}
\arguments{
\item{.tbl}{A dataset}

\item{.vars}{The variables to compute features on.}

\item{.var_nest1}{The variable \code{.var} is nested in.}

\item{.var_nest2}{The variable \code{.var_nest1} is nested in. This is a column
of \code{.tbl}.}

\item{features}{A list of functions (or lambda expressions) for the features to compute. \code{\link[fabletools:feature_set]{feature_set()}} is a useful helper for building sets of features.}

\item{...}{Additional arguments to be passed to each feature. These arguments will only be passed to features which use it in their formal arguments (\code{\link[base:formals]{base::formals()}}), and not via their \code{...}. While passing \code{na.rm = TRUE} to \code{\link[stats:cor]{stats::var()}} will work, it will not for \code{\link[base:mean]{base::mean()}} as its formals are \code{x} and \code{...}. To more precisely pass inputs to each function, you should use lambdas in the list of features (\code{~ mean(., na.rm = TRUE)}).}
}
\value{
An \code{smpspl_grid_features} object.
}
\description{
Features for sample splitting
}
