% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.default}
\alias{fit_model.model_spec}
\alias{fit_model.workflow}
\title{Wrapper to fit model}
\usage{
fit_model(.object, .data, ...)

\method{fit_model}{default}(.object, .data, ...)

\method{fit_model}{model_spec}(.object, .data, ...)

\method{fit_model}{workflow}(.object, .data, ...)
}
\arguments{
\item{.object}{an object}

\item{.data}{data to fit the model}

\item{...}{additional arguments for fitting the model.}
}
\description{
Convenient wrapper function to fit models of class \code{model_spec}, \code{workflow}
or \code{model_dfn}.
}
