% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_garch.R
\name{GARCH}
\alias{GARCH}
\title{GARCH Model}
\usage{
GARCH(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section).}

\item{...}{Further arguments for \code{\link[stats:arima]{stats::arima()}}}
}
\description{
Fits GARCH model using \link[fGarch:00fGarch-package]{fGarch}. Currently this is
fixed to be a GARCH(1, 1) model and the \code{formula} supplied is ignored.
}
\section{Specials}{

The \emph{specials} define the orders of the GARCH model. Currently this is fixed
to be a GARCH(1, 1) model and the formula supplied is ignored.
}

