% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innov.R
\name{innov}
\alias{innov}
\title{Simulates innovations}
\usage{
innov(nsim, dist = c("normal", "laplace", "digamma", "t", "stable"), ...)
}
\arguments{
\item{nsim}{number of innovations.}

\item{dist}{a character string identifying the distribution.}

\item{...}{parameters of the distribution.}
}
\description{
Simulates innovations
}
