% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_smpspl_reg.R
\name{is_smpspl_reg}
\alias{is_smpspl_reg}
\title{Check if model is registered}
\usage{
is_smpspl_reg(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{unused.}
}
\description{
Check if model is registered
}
