% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arima.R
\name{make_arima}
\alias{make_arima}
\title{Constructor for arima object}
\usage{
make_arima(phi, theta, delta)
}
\arguments{
\item{phi, theta}{numeric vectors of length \eqn{\ge 0} giving AR
    and MA parameters.}

\item{delta}{vector of differencing coefficients, so an ARMA model is fitted
to \code{y[t] - delta[1]*y[t-1] - ...}.}
}
\description{
Constructor for arima object
}
