% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-garch.R
\name{make_garch}
\alias{make_garch}
\title{Constructor for garch object}
\usage{
make_garch(omega, alpha, beta)
}
\arguments{
\item{omega}{numeric constant coefficient of the variance equation.}

\item{alpha}{numeric vector of autoregressive coefficients.}

\item{beta}{numeric vector of variance coefficients.}
}
\description{
Constructor for garch object
}
