% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspltest.R
\name{smpspltest}
\alias{smpspltest}
\title{Sum of squares}
\usage{
smpspltest(object, .metric, .max_lag)
}
\arguments{
\item{object}{an object.}

\item{.metric}{a character to filter the \code{metric} column.}

\item{.max_lag}{maximum lag}
}
\description{
Computes sum of squares of an object obtained from an application of
\code{fabletools::features}. This forms the chi-squared test statistic upto
scaling.
}
\examples{
\dontrun{
library(fable)
library(smpspl)
data <-
    tsibble::tsibble(x = rnorm(100), date = Sys.Date() + 0:99, index = date)

# Consider an AR(1) model
o <-
    ARIMA(x ~ pdq(1, 0, 0) + PDQ(0, 0, 0)) |>
    smpspl_grid(data, 2, 2)

# Calculate lanyard features
my_acf <-
    function(x) {
        data.frame(t = x, e = 0) |>
            lanyard::acf_metric(t, e) |>
            generics::tidy()
    }

#Calculate test statistic
o |>
 features(.resid, .subresid, .assessment, features = my_acf) |>
 tidyr::unnest(.nested_features) |>
 tidyr::unnest(.features) |>
 dplyr::filter(analysis_idx == 50, assessment_idx == 100) |>
 smpspltest("autocorrelation", 30)

#Compare with chi-square
stats::qchisq(0.95, df = 30) / 100
}
}
