% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl_list.R
\name{smpspl_list}
\alias{smpspl_list}
\title{Sample Split Wrapper for a List of Models}
\usage{
smpspl_list(model_list, fun, ...)
}
\arguments{
\item{model_list}{list of models to be fit.}

\item{fun}{a \link[smpspl:smpspl-package]{smpspl} function to be applied.}

\item{...}{parameters other than "object" of \code{fun}.}
}
\description{
Convenient wrapper for iterating over a list of models. The \code{object}
parameter of \code{fun} should take a registered model to be fit to the data.
}
